# AARCH64 some changes / performance increasing for Apple M1 / Cortex-A7x 