--- a/SFMT-neon.h
+++ b/SFMT-neon.h
@@ -10,10 +10,37 @@
 #ifndef SFMT_NEON_H
 #define SFMT_NEON_H
 
+#if defined(HAVE_SHA3) 
+#if !defined(__ARM_FEATURE_SHA3)
+/*
+ * SHA3 missing intrinsics
+ */
+__inline uint64x2_t veor3q_u64(uint64x2_t a, uint64x2_t b, uint64x2_t c)
+{
+    uint64x2_t r;
+    __asm__ ("eor3.16b %0,%1,%2,%3" :"=w"(r) :"w"(a), "w"(b), "w"(c));
+    return r;
+}
+__inline uint64x2_t vbcaxq_u64(uint64x2_t a, uint64x2_t b, uint64x2_t c)
+{
+    uint64x2_t r;
+    __asm__ ("bcax.16b %0,%1,%2,%3" :"=w"(r) :"w"(a), "w"(b), "w"(c));
+    return r;
+}
+#endif
+
+#define EOR3(a,b,c) veor3q_u64(a,b,c)
+#define BCAX(a,b,c) vbcaxq_u64(a,b,c)
+
+#else // NO SHA3
+/* emulate SHA3 */
+#define EOR3(a,b,c) veorq_u64(a, veorq_u64(b,c)) 
+#define BCAX(a,b,c) veorq_u64(a, vbicq_u64(b,c))
+#endif
+
+
 inline static void neon_recursion(uint32x4_t * r, uint32x4_t a, uint32x4_t b,
                                 uint32x4_t c, uint32x4_t d);
-
-
 
 /**
  * This function represents the recursion formula.
@@ -28,19 +55,15 @@ inline static void neon_recursion(uint32x4_t * r, uint32x4_t a, uint32x4_t b,
 {
     uint32x4_t v, x, y, z;
     static const uint32x4_t vzero = {0,0,0,0};
-    static const uint32x4_t vmask = {SFMT_MSK1, SFMT_MSK2, SFMT_MSK3, SFMT_MSK4};
-
+    static const uint32x4_t vmask_c = {~SFMT_MSK1, ~SFMT_MSK2, ~SFMT_MSK3, ~SFMT_MSK4};
 #define rotate_bytes(A, B, C) vreinterpretq_u32_u8(vextq_u8(vreinterpretq_u8_u32(A),vreinterpretq_u8_u32(B),(C)))
-
+    x = rotate_bytes(vzero, a, 16-SFMT_SL2);
     y = vshrq_n_u32(b, SFMT_SR1);
     z = rotate_bytes(c, vzero, SFMT_SR2);
+    z = EOR3(x, a, z);
     v = vshlq_n_u32(d, SFMT_SL1);
-    z = veorq_u32(z, a);
+    z = BCAX(z, y, vmask_c);
     z = veorq_u32(z, v);
-    x = rotate_bytes(vzero, a, 16-SFMT_SL2);
-    y = vandq_u32(y, vmask);
-    z = veorq_u32(z, x);
-    z = veorq_u32(z, y);
     *r = z;
 }